test unitario JobOpeningFinder creado junto con la implementacion del ApplicationService
