+++ src/Contexts/LLBE/Students/application/Find/StudentFinder.ts
@@ -2,7 +2,7 @@ import { StudentRepository } from '../../domain/StudentRepository';
 import { Student } from '../../domain/Student';
 import { StudentId } from '../../../Shared/domain/Students/StudentId';
 import { ApplicationService } from '../../../../Shared/domain/ApplicationService';
-import { StudentNotFound } from "../../domain/StudentNotFound";
+import { StudentNotFound } from '../../domain/StudentNotFound';
 
 export class StudentFinder extends ApplicationService {
   private repository: StudentRepository;
@@ -14,7 +14,9 @@ export class StudentFinder extends ApplicationService {
 
   async run(studentId: StudentId): Promise<Student> {
     const result = await this.repository.search(studentId);
-    if (result === null) { throw new StudentNotFound(`Not found a student with id ${studentId.value}`);}
+    if (result === null) {
+      throw new StudentNotFound(`Not found a student with id ${studentId.value}`);
+    }
     return result;
   }
 }
