+++ tests/Contexts/LLBE/JobOpenings/application/Update/JobOpeningUpgrader.test.ts
@@ -1,9 +1,9 @@
-import { JobOpeningMother } from '../domain/JobOpeningMother';
-import { JobOpeningRepositoryMock } from '../__mocks__/JobOpeningRepositoryMock';
-import { JobOpeningIdMother } from '../../Shared/domain/JobOpenings/JobOpeningIdMother';
+import { JobOpeningMother } from '../../domain/JobOpeningMother';
+import { JobOpeningRepositoryMock } from '../../__mocks__/JobOpeningRepositoryMock';
+import { JobOpeningIdMother } from '../../../Shared/domain/JobOpenings/JobOpeningIdMother';
 import { UpgradeJobOpeningRequestMother } from './UpgradeJobOpeningRequestMother';
-import { JobOpeningNotFound } from '../../../../../src/Contexts/LLBE/JobOpenings/domain/JobOpeningNotFound';
-import { JobOpeningUpgrader } from '../../../../../src/Contexts/LLBE/JobOpenings/application/JobOpeningUpgrader';
+import { JobOpeningNotFound } from '../../../../../../src/Contexts/LLBE/JobOpenings/domain/JobOpeningNotFound';
+import { JobOpeningUpgrader } from '../../../../../../src/Contexts/LLBE/JobOpenings/application/Update/JobOpeningUpgrader';
 
 let repository: JobOpeningRepositoryMock;
 let upgrader: JobOpeningUpgrader;
