+++ tests/Contexts/LLBE/JobOpenings/domain/JobOpening.test.ts
@@ -1,4 +1,4 @@
-import { CreateJobOpeningRequestMother } from '../application/CreateJobOpeningRequestMother';
+import { CreateJobOpeningRequestMother } from '../application/Create/CreateJobOpeningRequestMother';
 import { JobOpeningMother } from './JobOpeningMother';
 
 describe('JobOpening', () => {
